generalize web annotation patterns
