+++ readhomer_atlas/web_annotation/views.py
@@ -6,11 +6,15 @@ from django.urls import reverse_lazy
 from django.views.decorators.cache import cache_page
 
 from ..library.models import Node
+from .generators import (
+    WebAnnotationCollectionGenerator,
+    get_generator_for_kind,
+)
 from .shims import AlignmentsShim
 from .shortcuts import build_absolute_url
-from .utils import (
-    WebAnnotationCollectionGenerator,
-    WebAnnotationGenerator,
+
+
+from .utils import (  # WebAnnotationCollectionGenerator,; WebAnnotationGenerator,
     as_zero_based,
     preferred_folio_urn,
 )
@@ -24,7 +28,7 @@ def get_folio_obj(urn):
 
 
 @cache_page(settings.DEFAULT_HTTP_CACHE_DURATION)
-def serve_wa(request, urn, idx, format):
+def serve_wa(request, annotation_kind, urn, idx, format):
     # @@@ query alignments from Postgres
     alignment_by_idx = None
     alignments = AlignmentsShim(urn).get_alignment_data()
@@ -35,7 +39,8 @@ def serve_wa(request, urn, idx, format):
     if not alignment_by_idx:
         raise Http404
 
-    wa = WebAnnotationGenerator(urn, alignment)
+    generator_class = get_generator_for_kind(annotation_kind)
+    wa = generator_class(urn, alignment)
     if format == "text":
         return JsonResponse(data=wa.text_obj)
     elif format == "html":
@@ -45,20 +50,27 @@ def serve_wa(request, urn, idx, format):
 
 
 @cache_page(settings.DEFAULT_HTTP_CACHE_DURATION)
-def serve_web_annotation_collection(request, urn, format):
+def serve_web_annotation_collection(request, annotation_kind, urn, format):
     get_folio_obj(urn)
     # @@@ query alignments from Postgres
     alignments = AlignmentsShim(urn).get_alignment_data(fields=["idx"])
     paginator = Paginator(alignments, per_page=PAGE_SIZE)
     urls = {
-        "id": reverse_lazy("serve_web_annotation_collection", args=[urn, format]),
+        "id": reverse_lazy(
+            "serve_web_annotation_collection", args=[urn, annotation_kind, format]
+        ),
         "first": reverse_lazy(
             "serve_web_annotation_page",
-            args=[urn, format, as_zero_based(paginator.page_range[0])],
+            args=[urn, annotation_kind, format, as_zero_based(paginator.page_range[0])],
         ),
         "last": reverse_lazy(
             "serve_web_annotation_page",
-            args=[urn, format, as_zero_based(paginator.page_range[-1])],
+            args=[
+                urn,
+                annotation_kind,
+                format,
+                as_zero_based(paginator.page_range[-1]),
+            ],
         ),
     }
     data = {
@@ -74,7 +86,7 @@ def serve_web_annotation_collection(request, urn, format):
 
 
 @cache_page(settings.DEFAULT_HTTP_CACHE_DURATION)
-def serve_web_annotation_page(request, urn, format, zero_page_number):
+def serve_web_annotation_page(request, annotation_kind, urn, format, zero_page_number):
     get_folio_obj(urn)
 
     # @@@ query alignments from Postgres
@@ -86,12 +98,18 @@ def serve_web_annotation_page(request, urn, format, zero_page_number):
         page = paginator.page(page_number)
     except EmptyPage:
         raise Http404
-    collection = WebAnnotationCollectionGenerator(urn, page.object_list, format)
+    generator_class = get_generator_for_kind(annotation_kind)
+    collection = WebAnnotationCollectionGenerator(
+        generator_class, urn, page.object_list, format
+    )
     urls = {
         "id": reverse_lazy(
-            "serve_web_annotation_page", args=[urn, format, as_zero_based(page_number)]
+            "serve_web_annotation_page",
+            args=[urn, annotation_kind, format, as_zero_based(page_number)],
         ),
-        "part_of": reverse_lazy("serve_web_annotation_collection", args=[urn, format]),
+        "part_of": reverse_lazy(
+            "serve_web_annotation_collection", args=[urn, annotation_kind, format]
+        ),
     }
     data = {
         "@context": "http://www.w3.org/ns/anno.jsonld",
@@ -104,13 +122,18 @@ def serve_web_annotation_page(request, urn, format, zero_page_number):
     if page.has_previous():
         prev_url = reverse_lazy(
             "serve_web_annotation_page",
-            args=[urn, format, as_zero_based(page.previous_page_number())],
+            args=[
+                urn,
+                annotation_kind,
+                format,
+                as_zero_based(page.previous_page_number()),
+            ],
         )
         data["prev"] = build_absolute_url(prev_url)
     if page.has_next():
         next_url = reverse_lazy(
             "serve_web_annotation_page",
-            args=[urn, format, as_zero_based(page.next_page_number())],
+            args=[urn, annotation_kind, format, as_zero_based(page.next_page_number())],
         )
         data["next"] = build_absolute_url(next_url)
     return JsonResponse(data)
